fix(ui): Fix `<OrganizationTeams>` not listing all teams
This fixes a regression introduced in https://github.com/getsentry/sentry/pull/18067 when moving settings to lightweight org tree. With the original org details, *ALL* teams were included, with lightweight, we try to stub org.teams after org details is fetched during the transition. However, the endpoint we use for this is paginated, so this will break for orgs with many teams.

The solution is to replace usage of TeamsStore and `withTeams`, as these stores depend on the FULL teams list. By replacing this, we need to ensure that anywhere where we use teams, we need to make sure the UI supports a paginated version of the teams list. In this case, we have a search filter for the teams list, as well as limiting the number of teams returned on the initial load.